+++ src/compas_ui/rhino/artists/lineartist.py
@@ -1,0 +1,10 @@
+from __future__ import print_function
+from __future__ import absolute_import
+from __future__ import division
+
+from compas_ui.artists import PrimitiveArtist
+from compas_rhino.artists import LineArtist
+
+
+class LineArtist(LineArtist, PrimitiveArtist):
+    pass